+++ backend/rates/gecko.go
@@ -102,6 +102,7 @@ var (
 		"PLN": "pln",
 		"CZK": "czk",
 		"NGN": "ngn",
+		"ZAR": "zar",
 	}
 
 	// Copied from https://api.coingecko.com/api/v3/simple/supported_vs_currencies.
@@ -127,5 +128,6 @@ var (
 		"pln": "PLN",
 		"czk": "CZK",
 		"ngn": "NGN",
+		"zar": "ZAR",
 	}
 )