+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Display the wallet root fingerprint in the account info and device settings
 - Add support for Nigerian Naira (NGN)
+- Add support for South African Rand (ZAR)
 
 ## 4.38.0
 - Bundle BitBox02 firmware version v9.14.1